Move the shared OpenAI fixtures into conftest.py
Replaces the noqa suppressions from the previous commit, per review. The two
fixtures now live in conftest.py so pytest supplies them by discovery, which
means no test module imports or rebinds the names and Ruff F811 has nothing
to fire on.

mock_client and streaming_tool_call_chunks moved out of test_openai.py
unchanged, with the imports they need. Two chunk-type imports that only the
moved fixture used are dropped from test_openai.py.

posthog/test/ai is 420 passed, 88 skipped, the same as before the move.
