+++ posthog/test/ai/openai/test_openai.py
@@ -9,13 +9,11 @@ import pytest
 try:
     from openai.types.chat import ChatCompletion, ChatCompletionMessage
     from openai.types.chat.chat_completion import Choice
-    from openai.types.chat.chat_completion_chunk import ChatCompletionChunk
-    from openai.types.chat.chat_completion_chunk import Choice as ChoiceChunk
     from openai.types.chat.chat_completion_chunk import (
+        ChatCompletionChunk,
         ChoiceDelta,
-        ChoiceDeltaToolCall,
-        ChoiceDeltaToolCallFunction,
     )
+    from openai.types.chat.chat_completion_chunk import Choice as ChoiceChunk
     from openai.types.chat.chat_completion_message_tool_call import (
         ChatCompletionMessageToolCall,
         Function,
@@ -24,11 +22,11 @@ try:
     from openai.types.create_embedding_response import CreateEmbeddingResponse, Usage
     from openai.types.embedding import Embedding
     from openai.types.responses import (
+        ParsedResponse,
         Response,
+        ResponseFunctionToolCall,
         ResponseOutputMessage,
         ResponseOutputText,
-        ResponseFunctionToolCall,
-        ParsedResponse,
     )
     from openai.types.responses.parsed_response import (
         ParsedResponseOutputMessage,
@@ -48,13 +46,6 @@ except ImportError:
 pytestmark = pytest.mark.skipif(
     not OPENAI_AVAILABLE, reason="OpenAI package is not available"
 )
-
-
-@pytest.fixture
-def mock_client():
-    with patch("posthog.client.Client") as mock_client:
-        mock_client.privacy_mode = False
-        yield mock_client
 
 
 @pytest.fixture
@@ -246,106 +237,6 @@ def mock_openai_response_with_null_token_details():
             completion_tokens_details={"reasoning_tokens": None},
         ),
     )
-
-
-@pytest.fixture
-def streaming_tool_call_chunks():
-    return [
-        ChatCompletionChunk(
-            id="chunk1",
-            model="gpt-4",
-            object="chat.completion.chunk",
-            created=1234567890,
-            choices=[
-                ChoiceChunk(
-                    index=0,
-                    delta=ChoiceDelta(
-                        role="assistant",
-                        tool_calls=[
-                            ChoiceDeltaToolCall(
-                                index=0,
-                                id="call_abc123",
-                                type="function",
-                                function=ChoiceDeltaToolCallFunction(
-                                    name="get_weather",
-                                    arguments='{"location": "',
-                                ),
-                            )
-                        ],
-                    ),
-                    finish_reason=None,
-                )
-            ],
-        ),
-        ChatCompletionChunk(
-            id="chunk2",
-            model="gpt-4",
-            object="chat.completion.chunk",
-            created=1234567891,
-            choices=[
-                ChoiceChunk(
-                    index=0,
-                    delta=ChoiceDelta(
-                        tool_calls=[
-                            ChoiceDeltaToolCall(
-                                index=0,
-                                id="call_abc123",
-                                type="function",
-                                function=ChoiceDeltaToolCallFunction(
-                                    arguments='San Francisco"',
-                                ),
-                            )
-                        ],
-                    ),
-                    finish_reason=None,
-                )
-            ],
-        ),
-        ChatCompletionChunk(
-            id="chunk3",
-            model="gpt-4",
-            object="chat.completion.chunk",
-            created=1234567892,
-            choices=[
-                ChoiceChunk(
-                    index=0,
-                    delta=ChoiceDelta(
-                        tool_calls=[
-                            ChoiceDeltaToolCall(
-                                index=0,
-                                id="call_abc123",
-                                type="function",
-                                function=ChoiceDeltaToolCallFunction(
-                                    arguments=', "unit": "celsius"}',
-                                ),
-                            )
-                        ],
-                    ),
-                    finish_reason=None,
-                )
-            ],
-        ),
-        ChatCompletionChunk(
-            id="chunk4",
-            model="gpt-4",
-            object="chat.completion.chunk",
-            created=1234567893,
-            choices=[
-                ChoiceChunk(
-                    index=0,
-                    delta=ChoiceDelta(
-                        content="The weather in San Francisco is 15°C.",
-                    ),
-                    finish_reason=None,
-                )
-            ],
-            usage=CompletionUsage(
-                prompt_tokens=20,
-                completion_tokens=15,
-                total_tokens=35,
-            ),
-        ),
-    ]
 
 
 @pytest.fixture
